+++ client/src/cli.rs
@@ -81,12 +81,13 @@ pub struct CliOpts {
 	pub no_update: bool,
 	#[arg(long)]
 	pub p2p_client_restart_interval: Option<u64>,
-
 	/// AutoNAT behaviour mode: client, server, or disabled
 	#[arg(long, value_enum, default_value = "server")]
 	pub auto_nat_mode: ServiceMode,
-
 	/// Kademlia operation mode: client or server
 	#[arg(long, value_enum)]
 	pub operation_mode: Option<KademliaMode>,
+	#[arg(long, action = ArgAction::SetTrue)]
+	/// Flag that is set when the client is running as a local test mode. [should be only used for testing!]
+	pub local_test_mode: bool,
 }