+++ core/src/network/p2p/configuration.rs
@@ -262,6 +262,8 @@ pub struct LibP2PConfig {
 	/// Interval between requests in ping protocol (default: 60s).
 	#[serde(with = "duration_seconds_format")]
 	pub ping_interval: Duration,
+	/// Local test mode flag, currently only disables the private address filtering for KAD discovery and allows manual address confirmation of local addresses.
+	pub local_test_mode: bool,
 }
 
 impl Default for LibP2PConfig {
@@ -283,6 +285,7 @@ impl Default for LibP2PConfig {
 			bootstraps: vec![],
 			dht_parallelization_limit: 20,
 			ping_interval: Duration::from_secs(60),
+			local_test_mode: false,
 		}
 	}
 }